documentation and cleanup
