+++ src/payments-methods/classic/payment-methods.js
@@ -4,11 +4,21 @@ import { useEffect } from '@wordpress/element';
 import { renderAndInitAmazonCheckout } from '../../renderAmazonButton';
 import React from 'react';
 
+/**
+ * Returns the payment method's description.
+ *
+ * @returns {string}
+ */
 const Content = () => {
 	const settings = getBlocksConfiguration( 'amazon_payments_advanced_data' );
 	return decodeEntities( settings.description || '' );
 };
 
+/**
+ * Returns a react component and also sets an observer for the onCheckoutAfterProcessingWithSuccess event.
+ * @param {object} props
+ * @returns React component
+ */
 const AmazonPayBtn = ( props ) => {
 	useEffect( () => {
 		const unsubscribe = props.eventRegistration.onCheckoutAfterProcessingWithSuccess(
@@ -32,6 +42,12 @@ const AmazonPayBtn = ( props ) => {
 	return <div id="classic_pay_with_amazon" />;
 };
 
+/**
+ * Returns the Components that will be used by Amazon Pay "Classic".
+ *
+ * @param {object} props
+ * @returns React Component
+ */
 export const AmazonContent = ( props ) => {
 	return (
 		<React.Fragment>
